WorkWithEfitBook_images

+++ BookVisionWebApp/Services/Interfaces/IBookService.cs
@@ -5,7 +5,7 @@ namespace BookVisionWebApp.Services.Interfaces
     public interface IBookService
     {
         public Task<BaseResponce<IEnumerable<Book>>> GetAllBooks();
-        public Task<BaseResponce<Book>> GetBookById(int id);
+        public Task<BaseResponce<Book>> GetBookById(long id);
         public Task<BaseResponce<Book>> CreateBook(Book book);
         public Task DeleteBook(Book book);
         public Task<BaseResponce<Book>> EditBook(Book book);
